+++ packages/pangea-sdk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- AI Guard: `entities` from Redact is now an optional field.
+
 ## 6.10.0 - 2025-10-29
 
 ### Changed